requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dropbox==11.36.2
idna==3.6
ply==3.11
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
stone==3.3.1
urllib3==2.1.0
